refactor: 💡 Removed unwanted ./types/offer.ts
